+++ src/main/java/com/ceos/spring_vote_21st/vote/web/dto/request/CandidateCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.ceos.spring_vote_21st.vote.web.dto;
+package com.ceos.spring_vote_21st.vote.web.dto.request;
 
 import com.ceos.spring_vote_21st.member.domain.CeosTeam;
 import com.ceos.spring_vote_21st.vote.domain.Election;
@@ -8,8 +8,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class CandidateCreateRequestDTO {
-    private Long electionId;
-
     private String name;
 
     private CeosTeam team;
