+++ src/main/java/com/ceos/spring_vote_21st/vote/web/dto/request/VoteCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.ceos.spring_vote_21st.vote.web.dto;
+package com.ceos.spring_vote_21st.vote.web.dto.request;
 
 import com.ceos.spring_vote_21st.vote.domain.Vote;
 import lombok.Getter;
