+++ src/main/java/com/ceos/spring_vote_21st/vote/web/dto/response/CandidateWithVoteResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ceos.spring_vote_21st.vote.web.dto;
+package com.ceos.spring_vote_21st.vote.web.dto.response;
 
 
 import com.ceos.spring_vote_21st.vote.domain.Candidate;
@@ -9,16 +9,16 @@ import lombok.*;
 @NoArgsConstructor
 @Getter
 public class CandidateWithVoteResponseDTO {
-    private Long id;
-    private String name;
     private Long electionId;
+    private Long candidateId;
+    private String name;
     private int voteCount;
 
     public static CandidateWithVoteResponseDTO from(Candidate candidate, int voteCount) {
         return CandidateWithVoteResponseDTO.builder()
-                .id(candidate.getId())
-                .name(candidate.getName())
                 .electionId(candidate.getElection().getId())
+                .candidateId(candidate.getId())
+                .name(candidate.getName())
                 .voteCount(voteCount)
                 .build();
     }
